pyproject.toml
[project]
name = "johnson-dso"
version = "0.1.0"
description = ""
authors = [
    {name = "Ayslan Johnson",email = "ayslanjohnson@gmail.com"}
]
license = "MIT"
readme = "README.md"
package = [{ inlcude = "app" }]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
mypkg = ["*.txt", "*.rst"]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.8"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
requests = "^2.31.0"
httpx = "^0.26.0"
python-dotenv = "^1.0.1"
python-keycloak = "^5.3.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
httpx = "^0.26.0"
python-sonarqube-api = "^2.0.5"

[tool.poetry.scripts]
start = "uvicorn app.main:app --host 0.0.0.0 --port 8000 --reload"

[build-system]
requires = [
    "poetry-core>=2.0.0,<3.0.0",
    "setuptools>=60",
    "setuptools-scm>=8.0"
    ]
build-backend = "poetry.core.masonry.api"